refactor: layout portfolio token list
