+++ apps/wallet-mobile/src/features/Portfolio2/useCases/PortfolioTokensList/PortfolioWalletTokenList/PortfolioWalletTokenList.tsx
@@ -3,18 +3,18 @@ import BigNumber from 'bignumber.js'
 import * as React from 'react'
 import {FlatList, StyleSheet, Text, View} from 'react-native'
 
-import {Spacer} from '../../../../components'
-import {useSearch} from '../../../../Search/SearchContext'
-import {makeList} from '../../../../utils'
-import {Line} from '../../common/Line'
-import {TokenBalanceItem} from '../../common/TokenBalanceItem/TokenBalanceItem'
-import {TokenEmptyList} from '../../common/TokenEmptyList'
-import {TotalTokensValue} from '../../common/TotalTokensValue/TotalTokensValue'
-import {useGetPortfolioBalance} from '../../common/useGetPortfolioBalance'
-import {useGetTokensWithBalance} from '../../common/useGetTokensWithBalance'
-import {useStrings} from '../../common/useStrings'
+import {Spacer} from '../../../../../components'
+import {useSearch} from '../../../../../Search/SearchContext'
+import {makeList} from '../../../../../utils'
+import {Line} from '../../../common/Line'
+import {TokenEmptyList} from '../../../common/TokenEmptyList'
+import {useGetPortfolioBalance} from '../../../common/useGetPortfolioBalance'
+import {useGetTokensWithBalance} from '../../../common/useGetTokensWithBalance'
+import {useStrings} from '../../../common/useStrings'
+import {TotalTokensValue} from '../TotalTokensValue/TotalTokensValue'
+import {TokenBalanceItem} from './TokenBalanceItem'
 
-export const PortfolioWalletTokenScreen = () => {
+export const PortfolioWalletTokenList = () => {
   const strings = useStrings()
   const {styles} = useStyles()
   const {search, isSearching} = useSearch()
